+++ radix-engine-common/src/constants/mod.rs
@@ -1,13 +1,13 @@
 mod native_addresses;
 mod sbor_payload;
-mod scrypto_vm;
 mod transaction_construction;
 mod transaction_execution;
 mod transaction_validation;
+mod wasm;
 
 pub use native_addresses::*;
 pub use sbor_payload::*;
-pub use scrypto_vm::*;
 pub use transaction_construction::*;
 pub use transaction_execution::*;
 pub use transaction_validation::*;
+pub use wasm::*;
